+++ lib/spotify/song.ts
@@ -12,17 +12,3 @@ export interface ISong{
   duration: number;
   artists: Array<IArtist>;
 }
-
-export interface SongResponse{
-  id:string;
-  name: string;
-  href: string;
-  uri: string;
-  popularity: number;
-  duration_ms: number;
-  artists: Array<IArtist>;
-}
-
-export interface TopSongResponse{
-  items: SongResponse[];
-}
